User check in all traits.

+++ controllers/AdminDeleteTrait.php
@@ -12,26 +12,35 @@
 
 namespace base_core\controllers;
 
+use lithium\security\Auth;
 use lithium\g11n\Message;
 use li3_flash_message\extensions\storage\FlashMessage;
+use li3_access\security\AccessDeniedException;
 
 trait AdminDeleteTrait {
 
 	public function admin_delete() {
 		extract(Message::aliases());
+		$user = Auth::check('default');
 
 		$model = $this->_model;
 		$model::pdo()->beginTransaction();
 
 		$item = $model::find($this->request->id);
 
+		if ($user['role'] !== 'admin' && !$item->isOwner($user)) {
+			throw new AccessDeniedException();
+		}
+
 		if ($item->delete()) {
 			$model::pdo()->commit();
+
 			FlashMessage::write($t('Successfully deleted.', ['scope' => 'base_core']), [
 				'level' => 'success'
 			]);
 		} else {
 			$model::pdo()->rollback();
+
 			FlashMessage::write($t('Failed to delete.', ['scope' => 'base_core']), [
 				'level' => 'error'
 			]);
